Moved Quarkus profile implementation inside the test class.
